fix unmask removing service instead of keeping if unit file is symlink not pointing to devnull

+++ files/docker/systemctl.py
@@ -5036,7 +5036,12 @@ class Systemctl:
             _f = self._force and "-f" or ""
             logg.info("rm {_f} '{target}'".format(**locals()))
         if os.path.islink(target):
-            os.remove(target)
+            link_target = os.readlink(target)
+            if link_target == _dev_null:
+                os.remove(target)
+                logg.info("Unit %s was masked and has been unmasked.", unit)
+            else:
+                logg.info("Unit %s is not masked (symlink points to %s); nothing to do.")
             return True
         elif not os.path.exists(target):
             logg.debug("Symlink did not exist anymore: %s", target)
